+++ crates/lib/src/games/valve.rs
@@ -160,3 +160,9 @@ game_query_mod!(vrising, "V Rising", Engine::new(1_604_030), 27016);
 game_query_mod!(zps, "Zombie Panic: Source", Engine::new(17_500), 27015);
 game_query_mod!(moe, "Myth of Empires", Engine::new(1_371_580), 12888);
 game_query_mod!(mordhau, "Mordhau", Engine::new(629_760), 27015);
+game_query_mod!(
+    pvak2,
+    "Pirates, Vikings, and Knights II",
+    Engine::new(17570),
+    27015
+);